Isaac Sim 5.1 migration, USD fixes, and config updates
