+++ isaac_internals/apps/omni.isaac.sim.python.kit
@@ -556,8 +556,7 @@ enabled=true # Enable this for DLSS
 "omni.isaac.occupancy_map" = {}
 "omni.replicator.isaac" = {}
 "omni.kit.loop-isaac" = {}
-"omni.isaac.ros_bridge" = {}
-"omni.isaac.shapenet" = {}
+# Removed omni.isaac.ros_bridge and omni.isaac.shapenet as they are not available in Isaac Sim 5.1+
 "omni.isaac.synthetic_recorder" = {}
 
 #linux only extensions
