+++ simulator/grade_utils/__init__.py
@@ -1,0 +1,1 @@
+# Make grade_utils a package
